agregando mensajes de usuario/validaciones

+++ backend/src/controllers/categories.controllers.js
@@ -15,22 +15,35 @@ const fs = require("fs-extra");
 ctrlCategories.createCategory = async (req, res) => {
   const { name, description, subCategories } = req.body;
 
-  const result = await cloudinary.v2.uploader.upload(req.file.path);
+  if(name && description && subCategories){
+    const result = await cloudinary.v2.uploader.upload(req.file.path);
 
-  imgURL = result.url;
-  const newCategory = new Category({
-    name,
-    description,
-    subCategories,
-    imgURL,
-  });
+    imgURL = result.url;
+    const newCategory = new Category({
+      name,
+      description,
+      subCategories,
+      imgURL,
+    });
+  
+    try {
+      const categorySaved = await newCategory.save();
+  
+    //borra el archivo ya que ha sido subido a la nube
+    await fs.unlink(req.file.path);
+  
+    res.status(201).json({ ok: true, categorySaved });
+    } catch (error) {
+      res.status(425).json({
+        msg: error
+      })
+    }
+    
+  }else{
+    return res.json({ok:false})
+  }
 
-  const categorySaved = await newCategory.save();
-
-  //borra el archivo ya que ha sido subido a la nube
-  await fs.unlink(req.file.path);
-
-  res.status(201).json(categorySaved);
+  
 };
 
 ctrlCategories.getCategories = async (req, res) => {
@@ -40,7 +53,6 @@ ctrlCategories.getCategories = async (req, res) => {
 };
 
 ctrlCategories.updateCategoryById = async (req, res) => {
-  
   const updatedCategory = await Category.findByIdAndUpdate(
     req.params.categoryId,
     req.body,
@@ -51,7 +63,9 @@ ctrlCategories.updateCategoryById = async (req, res) => {
 
 //Funcion para borrar un elemento en base al id
 ctrlCategories.deleteCategoryById = async (req, res) => {
-  const deleteCategory = await Category.findByIdAndDelete(req.params.categoryId);
+  const deleteCategory = await Category.findByIdAndDelete(
+    req.params.categoryId
+  );
   res.status(200).json(deleteCategory);
 };
